moved models and views to a folder as new pattern and replace uncoding to coding in py files
