--- l10n_br_base/base_demo.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-	<data>
-
-	</data>
-
-</openerp>
