--- l10n_br_base/l10n_br_base_data.xml
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-
-    <data noupdate="True">
-
-        <function eval="('default',False,'country_id', [('res.partner', False)], ref('base.br'), True, False, False, False, True)" id="country_br_set" model="ir.values" name="set"/>
-
-        <!-- Default bank account description -->
-        <record id="fund" model="res.partner.bank.type">
-            <field name="name">Conta Corrente</field>
-            <field name="code">fund</field>
-        </record>
-
-        <record id="fund_acc_number_field" model="res.partner.bank.type.field">
-            <field name="name">acc_number</field>
-            <field eval="True" name="readonly"/>
-            <field name="bank_type_id" ref="fund"/>
-        </record>
-
-        <record id="fund_bank_field" model="res.partner.bank.type.field">
-            <field name="name">bank</field>
-            <field eval="True" name="readonly"/>
-            <field name="bank_type_id" ref="fund"/>
-        </record>
-
-        <record id="fund_contry_field" model="res.partner.bank.type.field">
-            <field name="name">country_id</field>
-            <field eval="True" name="readonly"/>
-            <field name="bank_type_id" ref="fund"/>
-        </record>
-
-        <record id="fund_acc_number_dig_field" model="res.partner.bank.type.field">
-            <field name="name">acc_number_dig</field>
-            <field eval="True" name="readonly"/>
-            <field name="bank_type_id" ref="fund"/>
-        </record>
-
-        <record id="fund_bra_number_field" model="res.partner.bank.type.field">
-            <field name="name">bra_number</field>
-            <field eval="True" name="readonly"/>
-            <field name="bank_type_id" ref="fund"/>
-        </record>
-
-        <record id="fund_bra_number_dig_field" model="res.partner.bank.type.field">
-            <field name="name">bra_number_dig</field>
-            <field eval="True" name="readonly"/>
-            <field name="bank_type_id" ref="fund"/>
-        </record>
-
-        <record id="bank_acc_bank_field" model="res.partner.bank.type.field">
-            <field name="name">bank</field>
-            <field eval="True" name="required"/>
-            <field name="bank_type_id" ref="base.bank_normal"/>
-        </record>
-
-        <record id="bank_acc_number_field" model="res.partner.bank.type.field">
-            <field name="name">acc_number</field>
-            <field eval="True" name="required"/>
-            <field name="bank_type_id" ref="base.bank_normal"/>
-        </record>
-
-        <record id="bank_acc_number_dig_field" model="res.partner.bank.type.field">
-            <field name="name">acc_number_dig</field>
-            <field eval="True" name="required"/>
-            <field name="bank_type_id" ref="base.bank_normal"/>
-        </record>
-
-        <record id="bank_bra_number_field" model="res.partner.bank.type.field">
-            <field name="name">bra_number</field>
-            <field eval="True" name="required"/>
-            <field name="bank_type_id" ref="base.bank_normal"/>
-        </record>
-
-        <record id="bank_bra_number_dig_field" model="res.partner.bank.type.field">
-            <field name="name">bra_number_dig</field>
-            <field eval="True" name="required"/>
-            <field name="bank_type_id" ref="base.bank_normal"/>
-        </record>
-
-	</data>
-
-</openerp>
