--- l10n_br_base/res_partner_view.xml
@@ -1,150 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-	<data>
-
-		<!-- Parceiro -->
-		<record id="view_l10n_br_base_partner_tree" model="ir.ui.view">
-			<field name="name">l10n_br_base.partner.filter</field>
-			<field name="model">res.partner</field>
-			<field name="inherit_id" ref="base.view_res_partner_filter" />
-			<field name="arch" type="xml">
-				<field name="name" position="after">
-					<field name="cnpj_cpf" />
-				</field>
-			</field>
-		</record>
-
-		<record id="view_l10n_br_base_partner_tree" model="ir.ui.view">
-			<field name="name">l10n_br_base.partner.tree</field>
-			<field name="model">res.partner</field>
-			<field name="inherit_id" ref="base.view_partner_tree" />
-			<field name="arch" type="xml">
-				<field name="display_name" position="after">
-					<field name="cnpj_cpf" />
-					<field name="inscr_est" />
-				</field>
-			</field>
-		</record>
-
-		<record id="view_l10n_br_base_partner_form" model="ir.ui.view">
-			<field name="name">l10n_br_base.partner.form</field>
-			<field name="model">res.partner</field>
-			<field name="inherit_id" ref="base.view_partner_form" />
-			<field name="arch" type="xml">
-				<field name="is_company" position="replace">
-					<field name="is_company" on_change="onchange_mask_cnpj_cpf(is_company, cnpj_cpf)" class="oe_inline" />
-				</field>
-				<field name="parent_id" position="before">
-					<group colspan="4">
-						<div class="oe_edit_only">
-							<label for="legal_name" string="Nome e Sobrenome" attrs="{'invisible': [('is_company','=', True)]}"/>
-							<label for="legal_name" string="Razão Social" attrs="{'invisible': [('is_company','!=', True)]}"/>
-						</div>
-							<field colspan="4" name="legal_name" nolabel="1"/>
-						<div class="oe_edit_only">
-							<label for="cnpj_cpf" string="CPF" attrs="{'invisible': [('is_company','=', True)]}"/>
-							<label for="cnpj_cpf" string="CNPJ" attrs="{'invisible': [('is_company','!=', True)]}"/>
-						</div>
-						<field colspan="4" name="cnpj_cpf" on_change="onchange_mask_cnpj_cpf(is_company, cnpj_cpf)" nolabel="1" />
-					</group>
-				</field>
-				<field name="street" position="replace">
-					<field name="street" />
-					<field name="number" placeholder="Número" />
-					<field name="district" placeholder="Bairro" />
-				</field>
-				<field name="street2" position="replace">
-					<field name="street2" placeholder="Complemento" />
-				</field>
-				<field name="state_id" position="attributes">
-					<attribute name="domain">[('country_id','=',country_id)]</attribute>
-				</field>
-				<field name="state_id" position="after">
-					<field name="l10n_br_city_id" placeholder="cidade" on_change="onchange_l10n_br_city_id(l10n_br_city_id)" />
-				</field>
-				<field name="city" position="replace">
-					<field name="city" invisible="1" />
-				</field>
-				<field name="zip" position="attributes">
-					<attribute name="on_change">onchange_mask_zip(zip)</attribute>
-					<attribute name="style">width: 50%%</attribute>
-				</field>
-				<notebook>
-					<page position="inside" string="Outros Dados">
-						<group string="Dados Fiscais">
-							<group>
-								<div>
-									<label for="inscr_est" string="Inscr. Estadual" attrs="{'invisible': [('is_company','!=', True)]}" />
-									<label for="inscr_est" string="RG" attrs="{'invisible': [('is_company','=', True)]}" />
-								</div>
-								<field name="inscr_est" nolabel="1"/>
-								<field name="inscr_mun" attrs="{'invisible': [('is_company','!=', True)]}"/>
-							</group>
-							<group>
-								<field name="suframa" attrs="{'invisible': [('is_company','!=', True)]}"/>
-							</group>
-						</group>
-					</page>
-				</notebook>
-				<xpath expr="//field[@name='child_ids']/form/sheet/group/div/field[@name='street']" position="after">
-					<field name="number" placeholder="Número"/>
-				</xpath>
-				<xpath expr="//field[@name='child_ids']/form/sheet/group/div/field[@name='street2']" position="attributes">
-					<attribute name="placeholder">Complemento</attribute>
-				</xpath>
-				<xpath expr="//field[@name='child_ids']/form/sheet/group/div/field[@name='street2']" position="after">
-					<field name="district" placeholder="Bairro"/>
-				</xpath>
-				<xpath expr="//field[@name='child_ids']/form/sheet/group/div/div/field[@name='state_id']" position="after">
-					<field name="l10n_br_city_id" placeholder="Cidade" on_change="onchange_l10n_br_city_id(l10n_br_city_id)" style="width: 40%%"/>
-				</xpath>
-				<xpath expr="//field[@name='child_ids']/form/sheet/group/div/div/field[@name='city']" position="attributes">
-					<attribute name="invisible">1</attribute>
-				</xpath>
-			</field>
-		</record>
-
-		<!-- res.partner.bank -->
-		<record id="view_l10n_br_base_partner_bank_form" model="ir.ui.view">
-			<field name="name">view_l10n_br_base.partner.bank.form</field>
-			<field name="model">res.partner.bank</field>
-			<field name="inherit_id" ref="base.view_partner_bank_form" />
-			<field name="arch" type="xml">
-				<field name="acc_number" position="replace">
-					<newline />
-					<field name="acc_number" select="1" />
-					<field name="acc_number_dig" select="1" />
-					<newline />
-					<field name="bra_number" select="1" />
-					<field name="bra_number_dig" select="1" />
-				</field>
-				<field name="street" position="after">
-					<field name="number" placeholder="Número..."/>
-					<field name="street2" placeholder="Complemento..."/>
-					<field name="district" placeholder="Bairro..."/>
-				</field>
-				<field name="state_id" position="after">
-					<field name="l10n_br_city_id" placeholder="Cidade" on_change="onchange_l10n_br_city_id(l10n_br_city_id)"/>
-				</field>
-				<field name="city" position="attributes">
-					<attribute name="invisible">1</attribute>
-				</field>
-			</field>
-		</record>
-
-		<record id="view_l10n_br_base_partner_bank_tree" model="ir.ui.view">
-			<field name="name">view_l10n_br_base.partner.bank.tree</field>
-			<field name="model">res.partner.bank</field>
-			<field name="inherit_id" ref="base.view_partner_bank_tree" />
-			<field name="arch" type="xml">
-				<field name="acc_number" position="replace">
-					<field name="acc_number" />
-					<field name="acc_number_dig" />
-					<field name="bra_number" />
-					<field name="bra_number_dig" />
-				</field>
-			</field>
-		</record>
-
-	</data>
-</openerp>
